refactor(#11): util -> utils 패키지명 수정
