+++ application/auth/src/test/java/io/jus/hopegaarden/HopeGaardenApplicationTests.java
@@ -1,10 +1,7 @@
 package io.jus.hopegaarden;
 
-import io.jus.hopegaarden.util.IntegrationHelper;
+import io.jus.hopegaarden.utils.IntegrationHelper;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.test.context.ActiveProfiles;
 
 class HopeGaardenApplicationTests extends IntegrationHelper {
